web: Refactor to use factory pattern
Use factory pattern to create the app instance.
This prevents the circular imports with database
models and makes the app simpler.
