+++ web/__init__.py
@@ -1,1 +1,3 @@
-from web.app import app
+from web.factory import create_app
+
+app = create_app(init_db=False)
